server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Khoi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.0"
fastapi = "^0.103.2"
pydantic = {extras = ["email"], version = "^2.4.2"}
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
discord-py = "^2.3.2"
beanie = "1.25.0"
boto3 = "^1.28.63"
pillow = "^10.0.1"
facebook-sdk = "^3.1.0"
pytz = "^2023.3.post1"
aiofiles = "^23.2.1"
aiohttp = "3.8.5"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
fastapi-discord = "^0.2.5"
tinify = "^1.6.0"
python-multipart = "^0.0.9"
orjson = "^3.10.3"
requests-oauthlib = "^2.0.0"

[tool.ruff]
line-length = 100

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403", "F405"]
"main.py" = ["F403"]
"scripts.py" = ["E711"]

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
